Move delegated types to a single file
