--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer.Protocol/DelegatedCompletionItemResolveParams.cs
@@ -1,12 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-using Microsoft.VisualStudio.LanguageServer.Protocol;
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Protocol
-{
-    internal record DelegatedCompletionItemResolveParams(
-        TextDocumentIdentifier HostDocument,
-        VSInternalCompletionItem CompletionItem,
-        RazorLanguageKind OriginatingKind);
-}
