--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer.Protocol/DelegatedCompletionResolutionContext.cs
@@ -1,7 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Protocol
-{
-    internal record DelegatedCompletionResolutionContext(DelegatedCompletionParams OriginalRequestParams, object? OriginalCompletionListData);
-}
