--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer.Protocol/DelegatedRenameParams.cs
@@ -1,9 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-using Microsoft.VisualStudio.LanguageServer.Protocol;
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Protocol
-{
-    internal record DelegatedRenameParams(VersionedTextDocumentIdentifier HostDocument, Position ProjectedPosition, RazorLanguageKind ProjectedKind, string NewName);
-}
